Qouraproject/qoura_be/requirements.txt
asgiref==3.7.2
astroid==2.15.6
backports.zoneinfo==0.2.1
black==23.7.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cloudinary==1.33.0
dill==0.3.7
dj-database-url==2.0.0
Django==4.2.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
flake8==6.1.0
gunicorn==21.2.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
psycopg2-binary==2.9.6
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==1.7.1
pylint==2.17.5
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
typing-extensions==4.7.1
urllib3==1.26.16
wrapt==1.15.0
